requirements.txt
requests
beautifulsoup4
python-telegram-bot
flask
PyDrive2
